pyproject.toml
[project]
name = "ai-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.119.0",
    "google-genai>=1.45.0",
    "inngest>=0.5.9",
    "llama-index-core>=0.14.5",
    "llama-index-readers-file>=0.5.4",
    "openai>=2.5.0",
    "python-dotenv>=1.1.1",
    "qdrant-client>=1.15.1",
    "streamlit>=1.50.0",
    "uvicorn>=0.38.0",
]
